% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_panoids.R
\name{sample_panoids}
\alias{sample_panoids}
\title{Draw random sample of GoogleStreetview Panoids}
\usage{
sample_panoids(sldf, mc.cores = 1, ...)
}
\arguments{
\item{sldf}{a SpatialLinesDataFrame object from which points are to be sampled.}

\item{mc.cores}{number of cores to use when drawing the sample. Defaults to 1.}

\item{...}{additional params to be passed to the spatialEco::sample.line.}
}
\value{
a SpatialPointsDataFrame object containing the following columns:
\itemize{
  \item{"panoid"}{Google Streetview PanoID obtained from Streetview metadata API.}
  \item{"lat"}{Latitude associated with the PanoID (returned by Streetview metadata API).}
  \item{"lon"}{Longitude associated with the PanoID (returned by Streetview metadata API).}
  \item{"month"}{Month during which the image associated with the PanoID was taken. Can be NA.}
  \item{"year"}{Year during which the image associated with the PanoID was taken. Can be NA.}
  \item{"sampled_lat"}{Latitude of sampled point (against which Streetview API was queried).}
  \item{"sampled_lon"}{Longitude of sampled point (against which Streetview API was queried).}
}
}
\description{
This function draws a random sample of GoogleStreetviewPanoids from a provided
SpatialLinesDataFrame object, usually constructed from the shapefile of a road
network. Spatial sampling is done via the spatialEco package, while interactions
with the Streetview API are handled by the robolyst/streetview github module.
No API keys are necessary for this function, as it only queries the Streetview
Metadata API.
}
