% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_panoids.R
\name{sample_panoids}
\alias{sample_panoids}
\title{Draw random sample of GoogleStreetview Panoids}
\usage{
sample_panoids(sldf, mc.cores = 1, ...)
}
\arguments{
\item{sldf}{a SpatialLinesDataFrame object from which points are to be sampled.}

\item{mc.cores}{number of cores to use when drawing the sample. Defaults to 1.}

\item{...}{additional params to be passed to the spatialEco::sample.line.}
}
\description{
This function draws a random sample of GoogleStreetviewPanoids from a provided
SpatialLinesDataFrame object, usually constructed from the shapefile of a road
network. Spatial sampling is done via the spatialEco package, while interactions
with the Streetview API are handled by the robolyst/streetview github module.
No API keys are necessary for this function, as it only queries the Streetview
Metadata API.
}
